cherry pick 022/04/07 회원테이블 수정 image -> oname, sname

+++ src/main/java/com/munhwa/prj/member/web/MemberController.java
@@ -1,20 +1,31 @@
 package com.munhwa.prj.member.web;
 
-import com.munhwa.prj.config.auth.LoginUser;
-import com.munhwa.prj.config.auth.dto.SessionUser;
-import com.munhwa.prj.member.service.MemberService;
-import com.munhwa.prj.member.vo.Auth;
-import com.munhwa.prj.member.vo.MemberVO;
-import com.munhwa.prj.news.service.NewsService;
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import com.munhwa.prj.common.entity.UploadFile;
+import com.munhwa.prj.common.entity.UploadFileVO;
+import com.munhwa.prj.common.service.FileUtils;
+import com.munhwa.prj.common.service.UploadFileService;
+import com.munhwa.prj.config.auth.LoginUser;
+import com.munhwa.prj.config.auth.dto.SessionUser;
+import com.munhwa.prj.member.service.MemberService;
+import com.munhwa.prj.member.vo.Auth;
+import com.munhwa.prj.member.vo.MemberVO;
+import com.munhwa.prj.news.service.NewsService;
+
+import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Controller
@@ -28,8 +39,12 @@ public class MemberController {
 
     @Autowired
     private PasswordEncoder passwordEncoder;
+    
+    @Autowired
+    private FileUtils fileUtils;
+    
 
-    // 마이페이지
+    // 마이페이지 (+ 새소식리스트)
     @PreAuthorize("hasRole('R01')")
     @GetMapping("/mypage.do")
     public String mypage(@LoginUser SessionUser sessionUser, Model model) {
@@ -47,6 +62,23 @@ public class MemberController {
     @GetMapping("/changeProfile.do")
     public String changeProfile() {
         return "changeProfile-member";
+    }
+    
+    // 프로필 업데이트
+    @PostMapping("updateProfile.do")
+    public String updateProfile(MemberVO vo, MultipartFile file) throws IOException {
+    	if (file != null) {
+    		UploadFile upload = fileUtils.storeFile(file);
+    		vo.setOname(upload.getOriginalFileName());
+    		vo.setSname(upload.getStoredFileName());
+    		
+    	}
+    	int n = memberDao.updateProfile(vo);
+    	if (n != 0) {
+    		return "redirect:memberChangeInfo.do";
+    	} else {
+    		return "error/404";
+    	}	
     }
 
     // 개인정보 변경 페이지
@@ -74,7 +106,9 @@ public class MemberController {
 
     // 비밀번호 업데이트
     @PostMapping("updatePassword.do")
-    public String updatePassword(MemberVO vo) {
+    public String updatePassword(MemberVO vo, String password1) {
+    	System.out.println(password1);
+    	vo.setPassword(passwordEncoder.encode(password1));
         int n = memberDao.updatePassword(vo);
         if (n != 0) {
             return "redirect:memberChangeInfo.do";
@@ -91,14 +125,26 @@ public class MemberController {
 
     // 회원탈퇴
     @PostMapping("/deleteMember.do")
-    public String deleteMember(MemberVO vo) {
-        int n = memberDao.deleteMember(vo);
+    public String deleteMember(RedirectAttributes attr, MemberVO vo , @LoginUser SessionUser user) {
+        int n = 0;
+    	if (passwordEncoder.matches(vo.getPassword(), user.getPassword()) == true) {
+        	n = memberDao.deleteMember(vo);  
+        	SecurityContextHolder.clearContext();
+        }    
         if (n != 0) {
             return "redirect:home.do";
         } else {
-            return "error/404";
+        	String message ="아이디 또는 비밀번호가 일치하지 않습니다.";
+        	attr.addFlashAttribute("message", message);
+            return "redirect:dropMember.do";
         }
     }
+    
+    @GetMapping("/test.do")
+    public String test() {
+        return "test-member";
+    }  
+
 
     // 회원가입폼
     @GetMapping("/signupForm.do")
